requirements.txt

viam-sdk==0.47.0

typing-extensions
streamdeck==0.9.7
